feat(open): open archive inputs in local mode (`archive` feature)
Factor the archive-extraction toolkit out of the db loader into a shared
`archive` module, and use it from `LocalLoader`: with the `archive` feature,
`quent-open local` extracts tar/zip archives — passed directly or found while
walking passed directories — into a scratch dir and scans them for telemetry,
alongside plain context directories. So a folder of downloaded run archives
opens like a folder of contexts.

- New `archive` Cargo feature (tar/flate2/zstd/zip/tempfile); `db` enables it.
  Default build is unchanged (no archive/codec deps).
- Same guards as before: reject entries that escape the dir or aren't plain
  files/dirs, cap bytes per-archive and run-wide, cap entry count, force
  extracted dirs owner-writable.
- `LocalLoader` gains a scratch `TempDir` and a `new()` constructor.
